Fix unit tests by splitting out table test Replace deprecated mockito verify method
